Moved class into it's own document. Made preventing duplicate additions better, cleaned up code and comments somewhat
channelCriteria.py
Created document and imported it into ManagePlaylist,py

ManagePlaylist.py
moved the check for videoID in the archive before videos are added to the json list. Removed the duplicate entry into the archive into PlaylistAdd,py using the add_to_playlist list as a stack and popping the results, using the popped id in the api call and also appending it to the archive. Also added a check for if they are in the list to be added that have yet to be added for some reason. Moving this check to the bottom allows for videos that meet all criteria to still affect the standardization process and threshold calculation. This makes the "popularity" of a video consistent.

PlaylistAdd.py
Renamed variables to make them more self-documenting
Removed the position snippet from the playlist_add_request - by default it will add videos to the end of the playlist (which can also be resorted from the actual playlist page on youtube)
Moved Archiving to this script. Popped values of the add_to_playlist list are appended to an archive list that is used to extend the current added_to_playlist list

Tacked on some cosmetic Input statements to the end of ManagePlaylis.py and PlaylistAdd.py to make it easier to troubleshoot / inspect upon completion.